Fix a bug in the method ConfigTreeNode::retain_deletion_nodes(...), where a deleted multi-node was not being retained on the deletion tree -- the multi-node, as well as its parents, were getting deleted. As a result, the xorpsh was not getting properly notified of the deleted multi-node via the "config_changed" callback.
The problem can be demonstrated by a rtrmgr template like:

system {
    name-server @: ipv4 {
        %help:   short "Configure domain name server";
        %delete:;
        /* foo: bool = false; */
    }
}

and xorpsh commands like:

create system name-server 1.1.1.1
commit
delete system name-server 1.1.1.1
commit

If there is a second xorpsh instance running, that instance won't
receive properly the deleted state:

pavlin@possum.icir.org# The configuration had been changed by user pavlin
DELTAS:
    system {
        name-server 1.1.1.1
    }
[edit]
pavlin@possum.icir.org#
pavlin@possum.icir.org# The configuration had been changed by user pavlin
[edit]
pavlin@possum.icir.org#
pavlin@possum.icir.org# show
    system {
        name-server 1.1.1.1
    }

[edit]
pavlin@possum.icir.org#

Note that if the "foo: bool = false;" node in the rtrmgr template
is uncommented, then the problem is masked.

Submitted by:	Marat Nepomnyashy


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8206 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/rtrmgr/conf_tree_node.cc
@@ -13,7 +13,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/rtrmgr/conf_tree_node.cc,v 1.107 2006/01/31 23:47:50 pavlin Exp $"
+#ident "$XORP: xorp/rtrmgr/conf_tree_node.cc,v 1.108 2006/02/27 18:48:30 pavlin Exp $"
 
 //#define DEBUG_LOGGING
 #include "rtrmgr_module.h"
@@ -1369,7 +1369,9 @@ ConfigTreeNode::retain_deletion_nodes(const ConfigTreeNode& them,
 	    my_child->delete_subtree_silently();
 	    found_deletion_child = true;
 	}
-	if (! found_deletion_child) {
+	if ((! found_deletion_child)
+	    && node_found
+	    && (! my_child->has_undeleted_children())) {
 	    // XXX: common child among both trees, hence just prune it
 	    my_child->delete_subtree_silently();
 	}
@@ -2214,6 +2216,20 @@ ConfigTreeNode::is_default_value(const string& test_value) const
 	return (false);
 
     return (test_value == _template_tree_node->default_str());
+}
+
+bool
+ConfigTreeNode::has_undeleted_children() const
+{
+    list<ConfigTreeNode *>::const_iterator iter;
+
+    for (iter = _children.begin(); iter != _children.end(); ++iter) {
+	const ConfigTreeNode* child = *iter;
+	if ((child != NULL) && (! child->deleted()))
+	    return (true);
+    }
+
+    return (false);
 }
 
 string
